PLAT-911 | Move ContentClassifier component
