+++ resources/assets/js/admin/store/mutations-types.js
@@ -1,6 +1,3 @@
-// Users
-export const USERS_SETUP_CURRENT = 'USERS_SETUP_CURRENT';
-
 // Lessons
 export const LESSONS_READY = 'LESSONS_READY';
 export const SETUP_LESSONS = 'SETUP_LESSONS';
@@ -10,7 +7,6 @@ export const SET_LESSONS_LOADING = 'SET_COURSE_STRUCTURE_LOADING';
 //Quiz questions
 export const SETUP_QUIZ_QUESTION = 'SETUP_QUIZ_QUESTION';
 export const UPDATE_QUIZ_QUESTION = 'UPDATE_QUIZ_QUESTION';
-export const SETUP_QUIZ_ANSWERS = 'SETUP_QUIZ_ANSWERS';
 export const CLEAR_QUIZ_QUESTION = 'CLEAR_QUIZ_QUESTION';
 
 // Flashcards
@@ -30,20 +26,3 @@ export const ADD_TAG = 'ADD_TAG';
 export const SETUP_GROUPS = 'SETUP_GROUPS';
 export const ADD_GROUP = 'ADD_GROUP';
 export const SET_GROUPS_LOADING = 'SET_COURSE_STRUCTURE_LOADING';
-
-// Nested set
-export const RESET_NESTED_SET_STATE = 'RESET_NESTED_SET_STATE';
-export const SET_NESTED_SET_LOADING = 'SET_NESTED_SET_LOADING';
-export const SET_NESTED_SET_SAVING = 'SET_NESTED_SET_SAVING';
-export const SETUP_NESTED_SET = 'SETUP_NESTED_SET';
-export const ADD_NESTED_SET_NODE = 'ADD_NESTED_SET_NODE';
-export const UPDATE_NESTED_SET_ORDER_NUMBERS = 'UPDATE_NESTED_SET_ORDER_NUMBERS';
-export const UPDATE_NESTED_SET_NODE = 'UPDATE_NESTED_SET_NODE';
-export const DELETE_NESTED_SET_NODE = 'DELETE_NESTED_SET_NODE';
-export const SET_NESTED_SET_EDITOR_MODE = 'SET_NESTED_SET_EDITOR_MODE';
-export const SELECT_NESTED_SET = 'SELECT_NESTED_SET';
-export const SET_EXPANDED_NESTED_SET = 'EXPAND_NESTED_SET';
-
-// Tags
-export const SET_TAXONOMIES_LOADING = 'SET_TAXONOMIES_LOADING';
-export const SETUP_TAXONOMIES = 'SETUP_TAXONOMIES';
