+++ resources/assets/js/admin/store/store.js
@@ -16,11 +16,11 @@ import alerts from 'js/store/modules/alerts';
 import flashcards from 'js/admin/store/modules/flashcards';
 import form from 'js/store/modules/form';
 import flashcardsSets from 'js/admin/store/modules/flashcardsSets';
-import taxonomyTerms from 'js/admin/store/modules/taxonomyTerms';
+import taxonomyTerms from 'js/store/modules/taxonomyTerms';
 import tags from 'js/admin/store/modules/tags';
 import courseStructure from 'js/admin/store/modules/courseStructure';
 import groups from './modules/groups';
-import taxonomies from 'js/admin/store/modules/taxonomies';
+import taxonomies from 'js/store/modules/taxonomies';
 
 Vue.use(Vuex);
