+++ resources/assets/js/components/global/contentClassifier/ContentClassifierEditor.vue
