+++ resources/assets/js/store/modules/taxonomyTerms.js
@@ -1,5 +1,4 @@
-import {NESTED_SET_EDITOR_MODES} from 'js/consts/nestedSet';
-import {nestedSetMutations, nestedSetGetters, nestedSetActions, initialState} from 'js/admin/store/modules/nestedSet';
+import {nestedSetMutations, nestedSetGetters, nestedSetActions, initialState} from 'js/store/modules/shared/nestedSet';
 import axios from 'axios';
 import { getApiUrl } from 'js/utils/env';
