+++ resources/assets/js/store/mutations-types.js
@@ -210,3 +210,20 @@ export const SOCKET_CONNECTION_RECONNECTED = 'SOCKET_CONNECTION_RECONNECTED';
 // Flashcards
 export const FLASHCARDS_SET_FLASHCARDS_SET = 'FLASHCARDS_SET_FLASHCARDS_SET';
 export const FLASHCARDS_UPDATE_FLASHCARD = 'FLASHCARDS_UPDATE_FLASHCARD';
+
+// Nested set
+export const RESET_NESTED_SET_STATE = 'RESET_NESTED_SET_STATE';
+export const SET_NESTED_SET_LOADING = 'SET_NESTED_SET_LOADING';
+export const SET_NESTED_SET_SAVING = 'SET_NESTED_SET_SAVING';
+export const SETUP_NESTED_SET = 'SETUP_NESTED_SET';
+export const ADD_NESTED_SET_NODE = 'ADD_NESTED_SET_NODE';
+export const UPDATE_NESTED_SET_ORDER_NUMBERS = 'UPDATE_NESTED_SET_ORDER_NUMBERS';
+export const UPDATE_NESTED_SET_NODE = 'UPDATE_NESTED_SET_NODE';
+export const DELETE_NESTED_SET_NODE = 'DELETE_NESTED_SET_NODE';
+export const SET_NESTED_SET_EDITOR_MODE = 'SET_NESTED_SET_EDITOR_MODE';
+export const SELECT_NESTED_SET = 'SELECT_NESTED_SET';
+export const SET_EXPANDED_NESTED_SET = 'EXPAND_NESTED_SET';
+
+// Tags
+export const SET_TAXONOMIES_LOADING = 'SET_TAXONOMIES_LOADING';
+export const SETUP_TAXONOMIES = 'SETUP_TAXONOMIES';
