+++ resources/assets/js/store/store.js
@@ -27,9 +27,8 @@ import tasks from 'js/store/modules/tasks';
 import flashcards from 'js/store/modules/flashcards';
 import siteWideMessages from 'js/store/modules/siteWideMessages';
 import form from 'js/store/modules/form';
-// TODO don't import from admin
-import taxonomies from 'js/admin/store/modules/taxonomies';
-import taxonomyTerms from 'js/admin/store/modules/taxonomyTerms';
+import taxonomies from 'js/store/modules/taxonomies';
+import taxonomyTerms from 'js/store/modules/taxonomyTerms';
 
 Vue.use(Vuex);
